requirement.txt
# Core ML dependencies
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0

# ONNX dependencies
onnx>=1.14.0
onnxruntime-gpu>=1.15.0

# TensorRT (install separately based on CUDA version)
# tensorrt>=8.6.0

# Server dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Client dependencies
openai>=1.0.0
aiohttp>=3.8.0

# Utility dependencies
PyYAML>=6.0
numpy>=1.24.0
packaging>=21.0

# Optional: for better logging and progress bars
tqdm>=4.65.0
rich>=13.0.0